requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.5
django-import-export==1.2.0
entrypoints==0.3
et-xmlfile==1.0.1
flake8==3.7.5
idna==2.8
jdcal==1.4
lxml==4.3.1
mccabe==0.6.1
odfpy==1.4.0
openpyxl==2.6.0
pkg-resources==0.0.0
pycodestyle==2.5.0
pyflakes==2.1.0
pytz==2018.9
PyYAML==3.13
requests==2.21.0
soupsieve==1.7.3
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.24.1
xlrd==1.2.0
xlwt==1.3.0
